backend/pyproject.toml
[project]
name = "historic-events-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Andre Birkus",email = "birkusandre@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "openai (>=1.66.3,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)"
]

[tool.poetry]
package-mode = false  # Disables package mode
packages = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
dev = "scripts:dev"